+++ internal/storage/v1/kafka/writer_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
 	"github.com/jaegertracing/jaeger/internal/metricstest"
-	"github.com/jaegertracing/jaeger/plugin/storage/kafka/mocks"
+	"github.com/jaegertracing/jaeger/internal/storage/v1/kafka/mocks"
 	"github.com/jaegertracing/jaeger/storage/spanstore"
 )
